+++ noodles-sam/CHANGELOG.md
@@ -34,6 +34,8 @@
 
   * sam/record/sequence: Implement `DerefMut`.
 
+  * sam/record/sequence/base: Implement `TryFrom<u8>`.
+
 ### Changed
 
   * sam: Rename `RecordExt` to `AlignmentRecord`.